requirements.txt
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.5
lxml==4.9.3